requirements.txt
flask-oidc @ git+https://github.com/puiterwijk/flask-oidc@e3fc54ee79e9656eb19c73829d925dd896c65c61
aiohttp
beautifulsoup4
Flask
anthropic
PyJWT
python-dotenv
textract
tiktoken
langchain
llama_index
tqdm
itsdangerous==2.0.1
Flask-Assets
attrs==23.1.0
flask
flask-cors
flask-restful
ghostscript
joblib
jwt
requests
flask[async]
gunicorn
requests
Flask-cors
aioboto3
asyncio
boto3
openai
numpy
scipy
tenacity
bs4
ratelimit
click
hrai_python
backoff
